# scolaire-front/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8092",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/sectors": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "operationId": "getAllSectors",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sectors"
        ],
        "operationId": "updateSector",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sectors"
        ],
        "operationId": "addKind",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/registrations": {
      "get": {
        "tags": [
          "Registrations"
        ],
        "operationId": "getAllRegistrations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistrationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Registrations"
        ],
        "operationId": "updateRegistration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Registrations"
        ],
        "operationId": "addRegistration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/programs": {
      "get": {
        "tags": [
          "Programs"
        ],
        "operationId": "getAllPrograms",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Programs"
        ],
        "operationId": "updateProgram",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Programs"
        ],
        "operationId": "addProgram",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          }
        }
      }
    },
    "/kinds": {
      "get": {
        "tags": [
          "Kinds"
        ],
        "operationId": "getAllKinds",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KindResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Kinds"
        ],
        "operationId": "updateKind",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KindResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Kinds"
        ],
        "operationId": "addKind_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KindResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/academieYears": {
      "get": {
        "tags": [
          "AcademieYears"
        ],
        "operationId": "getAllAcademieYears",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcademieYearResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AcademieYears"
        ],
        "operationId": "updateAcademieYear",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademieYearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademieYearResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AcademieYears"
        ],
        "operationId": "addKind_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademieYearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademieYearResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/registrations/{id}": {
      "get": {
        "tags": [
          "Registrations"
        ],
        "operationId": "getRegistrationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Registrations"
        ],
        "operationId": "deleteRegistrationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/registrations/by-page": {
      "get": {
        "tags": [
          "Registrations"
        ],
        "operationId": "listAllRegistrations",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseRegistrationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/programs/{id}": {
      "get": {
        "tags": [
          "Programs"
        ],
        "operationId": "getProgramById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Programs"
        ],
        "operationId": "deleteProgramById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/programs/by-page": {
      "get": {
        "tags": [
          "Programs"
        ],
        "operationId": "listAllPrograms",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseProgramResponse"
                }
              }
            }
          }
        }
      }
    },
    "/kinds/{id}": {
      "get": {
        "tags": [
          "Kinds"
        ],
        "operationId": "getKindById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KindResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Kinds"
        ],
        "operationId": "deleteKindById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/kinds/by-page": {
      "get": {
        "tags": [
          "Kinds"
        ],
        "operationId": "listAllKinds",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseKindResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SectorRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "SectorResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        }
      },
      "RegistrationRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "studentId": {
            "type": "integer",
            "format": "int64"
          },
          "administrativeAgentId": {
            "type": "integer",
            "format": "int64"
          },
          "classeId": {
            "type": "integer",
            "format": "int64"
          },
          "programId": {
            "type": "integer",
            "format": "int64"
          },
          "academieYearId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "academieYearId",
          "administrativeAgentId",
          "classeId",
          "date",
          "programId",
          "studentId"
        ]
      },
      "RegistrationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "studentId": {
            "type": "integer",
            "format": "int64"
          },
          "administrativeAgentId": {
            "type": "integer",
            "format": "int64"
          },
          "classeId": {
            "type": "integer",
            "format": "int64"
          },
          "programId": {
            "type": "integer",
            "format": "int64"
          },
          "academieYearId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ProgramRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "kindId": {
            "type": "integer",
            "format": "int64"
          },
          "archive": {
            "type": "boolean"
          }
        },
        "required": [
          "kindId",
          "name"
        ]
      },
      "ProgramResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "kindId": {
            "type": "integer",
            "format": "int64"
          },
          "archive": {
            "type": "boolean"
          }
        }
      },
      "KindRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        }
      },
      "KindResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        }
      },
...
      "AcademieYearRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "AcademieYearResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        }
      },
...
      "PageResponseRegistrationResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PageResponseProgramResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgramResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PageResponseKindResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KindResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}